Added new command: trace
The "test" command is now the same as calling "build", "run" and "trace" in succession. If you would prefer to only build and run, you can use the -notrace flag.

If you are using the -notrace flag, but would like to catch trace messages using a different method, you can use the "trace" method directly to receive log messages for the current project.

Currently support for webOS and Android

git-svn-id: https://nekonme.googlecode.com/svn/trunk@782 1509560c-5e2a-0410-865c-31c25e1cfdef