Added cross-compile support for Neko.
Working on Windows right now, will test on other platforms to confirm. 

Usage: If you pass "windows", "mac" or "linux" as the target, it should default to C++ if you are running that platform, otherwise it should default to Neko. This is the same if you call "neko" as the target, and pass a flag for the target you wish to use, like "neko -windows" or "neko -mac"

Neko should allow cross-compiling now, and ignore the run command for platforms other than the current. C++ should still only allow a compile for the current platform.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@981 1509560c-5e2a-0410-865c-31c25e1cfdef