Lib.exit will now close the process on Android, instead of minimizing
The default behavior of the back button on Android is to minimize the application, which I believe is the correct behavior. However, I think there should be a clean way to close the application, so it does a fresh restart the next time the user opens the application.

In my testing, Lib.close and Lib.exit would still resume where you left off (they call the same thing internally) and Lib.forceClose seemed to create issues the next time I tried to run the game (memory leak?). Sys.exit(0) seems to be the cleanest way of exiting... I'm not sure if it is identical to nme_terminate or if it is a little cleaner?

Anyway, I updated so Lib.exit will end the process, by default, if stage.onQuit is equal to close. This means that the default back button behavior is still to minimize, Lib.close will still minimize, and for Android specifically, Lib.exit will end the process, unless you defined a custom onQuit method.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1104 1509560c-5e2a-0410-865c-31c25e1cfdef