Added missing Int type to JNI return values, and improved the API for calling JNI methods
I like how simple it is to use a native C++ method after you call Lib.load. I wanted to introduce a similar method for interacting with Java methods, and fortunately I found a way to do it which should handle garbage collection for the JNI methods nicely, while also providing a simple API after calling JNI.createMemberMethod or JNI.createStaticMethod

The only thing that doesn't seem to work, which would be nice, is being able to use a function like JNI.createStaticMethod when you are defining the static value, like this:

private static var my_native_method = JNI.createStaticMethod("my.java.CustomClass", "customMethod", "()V");

I am not sure that the way I've implemented it creates a large cost in performance. If there is a problem, perhaps we could find a way to have the same behavior, at an API level, while optimizing at a code-level?

If it is a static method, you'll be able to call it like a regular function, like this:

my_custom_java_method ("Hello", "from", "Haxe!");

If it is a member method, it works the same way, but you pass the instance as the first value:

my_custom_member_method (jobject, "Hello", "World!");

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1117 1509560c-5e2a-0410-865c-31c25e1cfdef