Improved JNI to support two kinds of method calls.
By default, JNI.createMemberMethod and JNI.createStaticMethod will return a "traditional" method, which accepts as many parameters as the Java method it mirrors. However, (especially for dynamically generated code), you can now pass a true value for the fourth, "useArray" parameter to receive methods which accept only one parameter -- an array with all of the arguments.

You still need to pass the jmethod as the first argument when calling a member method

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1123 1509560c-5e2a-0410-865c-31c25e1cfdef