Started work on an SWC parser.
I am receiving an "invalid zip entry" error when it tries to parse the second entry (the XML file after the library SWF). When I ignored this error, it had a -5 zlib error (something to do with buffers). Perhaps there are some extra bytes somewhere in the file which causes the SWF entry to be read too short or too far?

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1173 1509560c-5e2a-0410-865c-31c25e1cfdef