Fixed child MovieClip names when creating a symbol from SWFs, and ignoring the Metadata and DefineSceneAndFrameLabelData tags.
Now when you create a named symbol from a SWF, the child clips will have their names from the Flash IDE, making it possible to use getChildByName to get pre-named child objects.

It appears that the Metadata tag is for defining the SWF size, frame rate and background color, which doesn't seem relevant, so I'm handling it as ignored. The data I found also suggested that the DefineSceneAndFrameLabelData is redundant, so I am also handling it as ignored.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1174 1509560c-5e2a-0410-865c-31c25e1cfdef