Improved Assets.getBitmapData so BitmapData instances are cached.
Because it is very common that you would want BitmapData instances to be shared, Assets.getBitmapData will now cache and reuse the same BitmapData instance automatically. If you would prefer to get a new instance each time, or avoid the cache, you can pass a false value as the second parameter.

In order to allow maintainance of the cache later in the application, the "cachedBitmapData" hash table is exposed, so you can clear individual values or the full table to help send instances to the garbage collector 

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1182 1509560c-5e2a-0410-865c-31c25e1cfdef