Update to the NMML specification. Added "spec.nmml" to illustrate how to use it.
As we move towards NME 3.2, I felt it was important to consider the NMML specification to make sure it is nice for the future. It will only ever become harder to make these kinds of changes, so now is as good of time as any.

The tools should be compatible with the previous format, but the full updated specification can be found in the new "spec.nmml" file, with documentation.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1188 1509560c-5e2a-0410-865c-31c25e1cfdef