Improved error output for the NME command-line tools
If you throw an error in Haxe, it reports the stack trace leading up to the error. In many cases this is the ideal behavior, but for our build tools, it is only confusing to have the stack that leads up to an unrelated error, like a bad path in your project file.

Alternatively, if we use trace or Lib.print, editors (like FlashDevelop) will not catch the message as an error message, so users would have to go to the output to find the log text that explained the error.

I have added a simple method in C, which lets us push messages (forcefully) to the error output stream. If this call is wrapped in a try...catch block, the message will still come through as an error, but the stack will be hidden.

If you compile using the -verbose option, the message will go out as a log message, then the real haxe error will be rethrown. In certain cases this should help us get additional information about an error, if necessary. However, in general this will only be firing when we know the cause of the error.

As before, if the user has an error in their code, the haxe output will come through and messages from the command-line tools will be suppressed (who needs to know "error while calling haxe.exe"?)

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1203 1509560c-5e2a-0410-865c-31c25e1cfdef