Added "VSDEBUG" value for debugging Windows applications using Visual Studio
I am not certain how well this works for debugging (as I haven't purposely thrown a crashing error) and I'm not sure if there's more that can be done (perhaps?) to set up an ideal debugging environment in Windows, but for now the command-line tools will look for a VSDEBUG value. This should generally be "devenv" or the path to your VCExpress.exe file. If you compile for Windows desktop, as debug, and have this flag set, it will open the file in Visual Studio, where you can hit play and (presumably) receive debugging.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1208 1509560c-5e2a-0410-865c-31c25e1cfdef