Made some more revisions to the NMML spec
I kept struggling to decide where the "main" attribute should go. Besides setting the main class, everything that has been in the <app /> node has basically been meta-data. Well, that and the "file" attribute, too, but I already looked at moving that out.

Today sealed it for me -- a first time developer assumed that the "package" attribute was the package for the main class, while "main" was the name of the class. There really wasn't anything to communicate otherwise, since he wasn't publishing for mobile.

I'm making a move to suggest <meta /> as the new standard tag for the title, description, package name, etc. I think this makes a lot of sense, since all of this data is really descriptive, and not as much functional.

With this in mind, I'm using <app /> for the "main" class, the output "file" and the build "path", as well as the "swf-version" when targeting Flash. Let me know if one or two of these feel misnamed or out-of-place.

So <meta /> for descriptive content, <app /> for more practical build properties and <window /> for how it should be rendered.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1223 1509560c-5e2a-0410-865c-31c25e1cfdef