Added a <swf /> tag which can embed SWF content and generate classes automatically
Usage:

I am not sure if this format will be final, but here is how it works right now.

In your NMML file, add <swf path="path/to/swf.swf" /> (you can also add an optional "rename" attribute). This will embed the SWF file as an asset in your project, and will generate Haxe classes for each of the classes found in the SWF.

To use this feature, you need to have "swf" installed and included in your project from haxelib. I have removed the SWF functionality from NME, and have moved it to a separate project, so it can develop and improve independent of NME releases.

So, if you have a symbol with the class "MySquare_Display", you will actually be able to call "new MySquare_Display ()" in your code. Similarly, if the class was "com.example.MySquare", you could import this class and use it in your code.

As IDEs grow to support "nme display" for completion, they will be able to find the generated classpath directories, and will be able to provide completion for these generated classes.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1256 1509560c-5e2a-0410-865c-31c25e1cfdef