Added a temporary patch to help prevent software-rendered shapes from disappearing
The issue seems to be related (still) to alpha masks in the software renderer. Once the left edge of an alpha run (I think) hits the left edge of the screen, the whole run would disappear. The patch I made earlier seems to have fixed the behavior for standard rectangles and bitmaps, but there is still an issue with more complex SWF-based shapes.

By increasing the visible_pixels when rendering a polygon, and by tweening with .snapping() to stay on even pixel positions, I can help extend when each horizontal line of the object will disappear. I'm still not sure of the exact cause

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1275 1509560c-5e2a-0410-865c-31c25e1cfdef