Added support for using both debug and release NDLLs for NME
Instead of publishing to the same file, now if you compile the NME NDLL with the -debug flag, it will append ".debug" before the file extension.

If you are building an NME application, the command-line tools will now search for a ".debug" file if it is not a release build. If it is not found, it will default back to the regular NDLL file. It will also copy the file to the same place as the original in the export directory, so there are no path issues to worry about there.

If it makes sense, the next release of NME can include debug and release builds of the NME NDLL for each platform, in order to make native debugging easier.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1365 1509560c-5e2a-0410-865c-31c25e1cfdef