Beginning merge of Jeash 0.8.7 classes into NME
There is a lot of similarity between the NME's Haxe API and Jeash, both at a high level and in their private method architecture. Instead of trying to separately maintain parity, there is a lot of sense in working together.

Although this initial import has a lot of redundant code, it will be possible to blend classes much more fully, especially if the rendering for Jeash is refactored a little bit.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1374 1509560c-5e2a-0410-865c-31c25e1cfdef