Helped hide internal methods and properties from code completion
When you receive code completion from the haxe compiler, it exposes private objects if you are extending that object. In other words, if they can be accessed. This is the right behavior, but based on the design of NME, it seems best to flag these as /** @private */, which (in my addin for MonoDevelop, at least) forces an object to be hidden from completion.

For example, since most objects extend Sprite, you would see "nmeAddChild", "nmeBroadcast" and a host of other internal-only methods. If you were to see your own private members, it seems like a fine behavior, but NME's internals are "even lower level" than an average framework, and should be hidden from users in my mind.

For now I have left the static CFFI calls exposed. If you extend a class, then look for a static, you will see these methods. Although they normally should not be called directly, it seems fair that an advanced user may consider (at their own risk) calling these directly. Since they are not in the path of normal use (like completion for "this.") I think it is safe to leave them unmasked for the time being.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1503 1509560c-5e2a-0410-865c-31c25e1cfdef