User-defines ("-D") will pass onto the Haxe compile process as well
There is still a distinction between defines (internal to the command-line tools) and compileFlags (passed onto the Haxe compile process).

However, sometimes you want defines to pass through into Haxe. For example, defining "fdb" from the command-line. Now, when you specify a define against the command-line tools (like "nme test project.nmml flash -Dfdb") it will define this value in the command-line tools, and it will pass it onto the Haxe compiler.

This still allows flexibility and distinction within the NMML format, but provides an easier way to push values to the Haxe compiler, when necessary.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1522 1509560c-5e2a-0410-865c-31c25e1cfdef