Including a "public static function main" is now optional
As before, if your main class as a "public static function main," it will be called when your application starts. However, if this does not exist, the project template will now assume that you class should be instantiated and added as a child of Lib.current, like this:

Lib.current.addChild (new YourMainClass ());

This should reduce the overheard for creating new projects, and help reduce the learning curve for new users.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1551 1509560c-5e2a-0410-865c-31c25e1cfdef