Changed the path and name of the ".build" file
Instead of saving to a hidden file in the output directory, like "Export/.build", the file will be saved next to the NMML file for the project, like "TestProject.build"

It feels best not to modify the NMML file when generating new builds, but this number should stay constant over the lifetime of the project, so it should not be blown away when cleaning the output directory, and would be best to commit to version control. Since the previous file was hidden on Unix systems, it was ignored by default by version control.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1567 1509560c-5e2a-0410-865c-31c25e1cfdef