Use console when creating a Windows debug application
Ideally, I would like to be able to include AttachConsole() in the WinMain method of HXCPP. This would make it possible to redirect output whenever an application has a console available.

Instead of showing a command-prompt window in Explorer, it would always launch cleanly. However, from a command-prompt or IDE, the trace messages would always come through.

The problem with using AttachConsole is that the WinMain of HXCPP is created using a macro define, which cannot have #include <Windows.h> ... for now, this should help improve the experience of debugging NME applications on Windows.

You can define SHOW_CONSOLE or HIDE_CONSOLE if you would like to always hide or show the command-prompt window. Otherwise, it will use the console subsystem when debugging, or the windows subsystem on release.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@1667 1509560c-5e2a-0410-865c-31c25e1cfdef