Implemented initial support for both GL1 and GL2 features, using a Haxe-based flag rather than setting a define when compiling the native platform libraries.
You can access GL2 features by using <window shaders="true" /> in your NMML. This is probably temporary, and should only be operable for desktop targets. There are a couple OpenGL 2 calls that are not present in OpenGL ES 2.0, so if you try to enable shader support on mobile platforms it should crash at runtime.

This is currently a development-only feature.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@2028 1509560c-5e2a-0410-865c-31c25e1cfdef