Working on new command-line tools
If you "nme rebuild tools" or any of the ordinary NME commands, it will still build and run the old tools. However, if you add "-new" to rebuild or any other command, it will redirect to the new tools.

Much of the architecture is in place, and features are being migrated over. The tools are currently stable for BlackBerry and the Mac desktop.

Rather than using a monolithic "installer" class, the new system is project driven. Once a project is defined, all of the build tools read and respond to the project file.

The NMEProject class can merge itself with other instances, and has already helped remove a lot of the cruft and guesswork when it comes to how a project is going to be built.

There is an NMMLParser class that extends the NMEProject class with logic to handle NMML files, but this is optional. We're headed down a path where project files could be based in Haxe code, and skip NMML entirely.

The primary benefits of a Haxe-based project file, apart from a lot more control over logic, code completion and error checking, will be the inclusion of additional tools at build time. Eventually this should support dynamically generated assets as a part of the project file. This opens the door to pre-processing image, SWF, sound and other asset files.

If none of this sounds interesting, don't worry -- NMML should continue to work. The new tools are still a work-in-progress, but it will get there.

One other focus with the new tools has been reusing functionality through "helper" classes and better "-verbose" output, to get a better idea of what's going on with the tools without looking at source code or guessing.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@2125 1509560c-5e2a-0410-865c-31c25e1cfdef