Disabling LZMA for native targets other than Windows, since the library appears to be Windows specific? (HANDLE, LPVOID, CRITICAL_SECTION...)
Long term, if this were added it would also be nice if the code were moved to its own file, like "Lzma.cpp", which could then implement the body of the nme_lzma methods. We would probably also want to move the code to sdl-static, but first I'm not sure if it would be difficult to make this compatible with more targets? It may be a difference between Win32 and POSIX threading.

git-svn-id: https://nekonme.googlecode.com/svn/trunk@2165 1509560c-5e2a-0410-865c-31c25e1cfdef