Adding an initial NME extension API, starting with "nme_register_bps_event_handler"
The majority of the functionality in an NME native extension can be done in Haxe, or in C/C++/Objective-C/Java code that is independent of the NME core.

However, it may make sense to have headers for a lightweight native API, when it would make sense for an extension to make low-level calls. For example, system events on BlackBerry drive functionality like BBM, Scoreloop, sensors and other services. Although sensors may make sense in the core of NME, it is probably better to have services like BBM and Scoreloop as extensions to NME. Since the services drive from the event loop, an extension would need a way to hook in at the native level.

Currently the idea is that an extension would include "NME\extension.h" to gain access to this API. Perhaps we might provide some help for resolving the paths and default properties for a Build.xml file down the road.



git-svn-id: https://nekonme.googlecode.com/svn/trunk@2179 1509560c-5e2a-0410-865c-31c25e1cfdef