Big improvements to the way assets and libraries are handled.
Initial improvements to assets so the format can be separated from the source file extension, and so assets can be defined only in memory rather than having to exist on the file system first.

The SWF class generation has been disabled, in favor of integrated support in the Asset.getBitmap and new Asset.getMovieClip methods. In order to access symbols from a library, the symbol name is prefixed by the name of the library and a colon. For example, if you include "library.swf", you can use "library:bg1.jpg" or whatever the exported name is for the symbol. The <library /> node accepts both a path and an optional name, if you need to control the reference name for the asset library

There is now initial support for XFL libraries, as well as SWF. The Assets class is now transparently handling both SWF and XFL libraries using the same API. Although the SWF format has limitations for HTML5, the XFL format should not. Currently the processing occurs in the NME tools, then the result is unserialized at runtime, which should theoretically keep file sizes dramatically down (since XFL tends to be large due to the XML) while also working across all targets.

Continued work will go forward in the future

git-svn-id: https://nekonme.googlecode.com/svn/trunk@2210 1509560c-5e2a-0410-865c-31c25e1cfdef