Move to "nmedev" haxelib
If "SDL_STATIC" is defined, it will use that path in the project Build.xml file, otherwise it will look for the "nmedev" haxelib, which includes static libraries and headers.

If you do not have "command-line.n" or the NDLL for your platform, it should search for it in "nmedev" as a backup. This can make it possible for developers to check out NME from the source, and begin improving it without compiling these files themselves, and without including them in the NME repository.