Confirm that "nme_dev" is being defined properly
Haxe 3 supports values for defines. Now there will be an "nme_dev" value with the current version, similar to the new "haxe_dev" value

If you are maintaining Haxe 2 compatibility, you can adapt to NME 3.6 by either checking for "nme_dev" existing, or if using "haxe3", since no older version of NME supports Haxe 3.

With Haxe 3, you will be able to perform comparisons, like...

#if (nme_ver < "3.6.1")

#if (nme_ver > "3.6.0")

#if (nme_ver == "3.6.0")

In order for comparisons to work properly, you must surround it with parenthesis, and you must compare as a string, since "3.6.0" is not a valid number.