Revert "remove extra ignore files someone may have pushed accidentally"
This reverts commit e707a9f47d665da73af59b8ae5ed6df80b31383f.

When you clone the repository from GIT, some of these directories are empty. Unless the folder for your current platform is present, "haxelib path nme" throws strange errors. These files are added to ensure the directories are present, and to avoid the confusing "Lib nme version dev does not have a neko dll for your system" and subsequent problems it generates :)