Define "haxe3" in NMML, for easier referencing if including version-specific Haxe flags
Removed the no-pattern-matching and log defines. The next version of NME is a clean break, targeting only Haxe 3. Although these defines are probably convenient for porting older libraries easier, I think this should be (at best) a project-specific decision for the developer, or (at worst) added in an include.nmml for the library.

The big shift for the switch() statement is instead of MyEnum.VALUE, you use VALUE in the case statements. The second is that only static runtime values appear to be eligible for a Haxe 3 switch, so sometimes it may make sense to use if() statements instead. This can cause bumps, but I think this needs to be up to the user to disable or not.