Added the SCF sound manager to nme.media
The soundmanager wraps and abstracts the sound API, branching to conform
to the android target specifically, while deferring to the standard API
for every other target. It also reads sound lengths (from wavs only)
from a text file and maps them to sound paths. This needs some work but
could conceivably be bundled with NME when tested and integrated.

Generating sound length data could conceivably be done with macros,
though bundling a binary format generated with the tool we built for SCF
is probably the best "simple stupid" solution.

Also added nme.utils.Property, which is just a generic class for
wrapping a value and tracking changes to it.