Replaced Sound.java
The current android template still uses methods proved through rigorous
testing to be dysfunctional, whereas our (currently outdated) Sound
manager for SCF remains a stable design. I simply dumped ours into the
current repository, which will most certainly break android builds
currently: See that as a kick in the pants to get it working ;-)

nme.media.soundmanager.io.android.AndroidAudio and
nme.media.soundmanager.io.android.SoundPool make hefty use of JNI calls
to the template.

I want to fix Android sound properly, but as far as I can tell the only
way to do so is to embrace what it actually is; Not compatible with the
Flash Sound API design. My hope for the future is to have an API that is
fast and loose enough that you can prototype quickly, but have a rigid,
dependable API to fall back on when the time comes to ship a product
that doesn't blow up.