Added "allow trans" flag to SimpleSurface
Internally, SimpleSurface may or may not have alpha. Depending on how the
surface was loaded or whether you have tried to perform an operation that
requires alpha, it will update automatically. This is a good balance for
performance and memory.

Now, the "allow trans" flag will mask incoming "setPixel" and "setPixels"
calls on the when the surface is supposed to remain opaque.

This should fix operations such as:

var bitmapData = new BitmapData (100, 100, false);
bitmapData.setPixel32 (0, 0, 0x00FFFFFF);
trace (bitmapData.getPixel32 (0, 0));