Windows ICO: Add 8-bit images
Some icons where not displaying correctly. For example, when you place the cursor over the taskbar icon and a "preview" window appears, the default icon was shown instead of the game icon. Looking into various icons of Windows applications, I have found that they include the 8-bit versions of at least the sizes of 48, 32 and 16. Furthermore, none of them used the "96" size, so I removed it from the 32-bit versions. Also, they are written from larger to smaller (just changed the order in the array.  A warn message should be shown if the input image has more than 255 colors.  In that case we look for the nearest color rhat is already in the palette.  The algorithm may be improved, but actually is better design to use less than 255 colors anyway...  Now we have reached a "standard" windows icon format.