Display stats: show and verbose levels
DisplayStats are enabled on debug or when -DNME_DISPLAY_STATS is set.
The following keys are used to display it on screen:
Alt+1 : toggle visibility
Alt+2 : change verbose level
(on mac, should be CMD instead of Alt)

If you use the flag  -DNME_DISPLAY_STATS, the stats are shown from
launch. The flag also enables DisplayStats on release mode.

To set an initial verbose level use
-DNME_DISPLAY_STATS=1 //level 1
-DNME_DISPLAY_STATS=2  //level 2