GameController check axis value in pairs
Receives Axis event on pairs instead of individual. Check for the x and
y values, or the axis array (changed values in axis array will be "e.id"
and "e.id+1" as e.id will be always even number).
Therefore, the "value" variable is removed.  This should reduce the
number of events send from C++ to Haxe and it should be the intended
initial behaivor.