Reverted CONTROLLER events, add mapping for legacy compatibility
It works with
https://github.com/HaxeFlixel/flixel-demos/tree/dev/Input/GamepadTest

,and also with the included HandlingJoystickEvents, using the joystick
API.

The difference is that we treat all the gamecontrollers as "XInput" type
(in legacy, diferent gamecontroller types where retrieved). There is a
mapping from SDL GAMECONTROLLER API to the XINPUT values but it can be
disabled removing openfl_legacy define.

The GameInput is also implemented but I haven't test it much. I changed
the joystick Dynamic objects to a simple Index for optimization since we
are already using JoystickEvents instead of Lime's GamePad object.