Fix "Failed to make complete framebuffer object" crash on iOS 13
Almost every time, when i minimize application on the iPhone with installed iOS 13 i get this crash "Failed to make complete framebuffer object". And this tread on Apple forum has advice on how to fix it: https://forums.developer.apple.com/thread/23864
I've implemented it and it helped me to fix this crash. I've also checked if it doesn't break anything on older iOS versions and it doens't